requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.1
controlnet-aux>=0.0.6
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.21.0
matplotlib>=3.7.0
tqdm>=4.65.0
python-dotenv>=1.0.0
huggingface-hub>=0.16.0
